requirements.txt
pandas==1.1.3
numpy==1.19.2
flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.7.0